Rename Nabla file to DifferentialOperators
